requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.13.3
pydantic-settings==2.6.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
redis==5.1.1
fakeredis==2.26.1
email-validator==2.2.0
pytest==8.3.3
httpx==0.27.2
ruff==0.6.9
